new questions added to "others" category
adding the newly added questions selected by the admin will ba saved to the database letting the client use that question

+++ launcher.py
@@ -14,19 +14,18 @@ def get_token(q: multiprocessing.Queue) -> None:
         q.put(request.args["token"])
         return Response("", 204)
 
-    eventlet_socket = eventlet.listen((HOST, PORT))
-    eventlet.wsgi.server(eventlet_socket, application)
     run_simple(HOST, PORT, application, use_reloader=True, use_debugger=True, use_evalex=True,
-               reloader_type='watchdog', threaded=True)
+               reloader_type='watchdog')
 
 
 if __name__ == "__main__":
-    eventlet_socket = eventlet.listen((HOST, PORT))
-    eventlet.wsgi.server(eventlet_socket, application)
     q = multiprocessing.Queue()
     p = multiprocessing.Process(target=get_token, args=(q,))
+    eventlet_socket = eventlet.listen((HOST, PORT))
+    eventlet.wsgi.server(eventlet_socket, application)
     p.start()
     print("waiting")
     token = q.get(block=True)
-    p.terminate()
+    if p.is_alive():
+        p.terminate()
     print(token)
